Split blas.py into a blas/ package
